requirements.txt
# AI-Powered Resume Screening System
# Python Package Requirements

# Core dependencies
# spacy>=3.7.0  # OPTIONAL - uncomment if you can install it with: python -m spacy download en_core_web_sm
PyPDF2>=3.0.0
python-docx>=1.1.0
pandas>=2.1.0
numpy>=1.24.0

# Web UI
streamlit>=1.28.0

# NLP and ML
nltk>=3.8
scikit-learn>=1.3.0

# Utilities
python-dateutil>=2.8.0
reportlab>=4.0.0  # For PDF generation

# Optional but recommended
openpyxl>=3.1.0  # For Excel export
matplotlib>=3.8.0  # For visualizations
seaborn>=0.13.0  # For visualizations
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
opencv-python>=4.8.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.10.0
flake8>=6.1.0
